Enabled cluster tests in the following files,
1. python/ray/tests/test_failure_2.py
2. python/ray/tests/test_get_locations.py
3. python/ray/tests/test_logging.py
4. python/ray/tests/test_metrics_agent.py
5. python/ray/tests/test_multi_node_2.py
6. python/ray/tests/test_multinode_failures.py
7. python/ray/tests/test_multinode_failures_2.py
8. python/ray/tests/test_node_manager.py
9. python/ray/tests/test_object_manager.py
10. python/ray/tests/test_object_manager_2.py
11. python/ray/tests/test_output.py
