+++ python/ray/tests/test_multi_node_2.py
@@ -332,21 +332,21 @@ def test_detached_actor_autoscaling(ray_start_cluster_head):
     assert len(down) == 0
 
 
-def test_multi_node_pgs(ray_start_cluster):
-    cluster = ray_start_cluster
+def test_multi_node_pgs(ray_start_cluster_enabled):
+    cluster = ray_start_cluster_enabled
     cluster.add_node(num_cpus=2)
     cluster.wait_for_nodes(2)
     ray.init(address=cluster.address)
 
     pgs = [ray.util.placement_group([{"CPU": 1}]) for _ in range(4)]
 
-    ready, not_ready = ray.wait([pg.ready() for pg in pgs], timeout=5, num_returns=4)
+    ready, not_ready = ray.wait([pg.ready() for pg in pgs], timeout=20, num_returns=4)
     assert len(ready) == 2
     assert len(not_ready) == 2
 
     cluster.add_node(num_cpus=2)
     cluster.wait_for_nodes(3)
-    ready, not_ready = ray.wait([pg.ready() for pg in pgs], timeout=5, num_returns=4)
+    ready, not_ready = ray.wait([pg.ready() for pg in pgs], timeout=20, num_returns=4)
     assert len(ready) == 4
     assert len(not_ready) == 0
 
@@ -356,7 +356,7 @@ def test_multi_node_pgs(ray_start_cluster):
         print(".")
         more_pgs = [ray.util.placement_group([{"CPU": 1}]) for _ in range(2)]
         ready, not_ready = ray.wait(
-            [pg.ready() for pg in more_pgs], timeout=5, num_returns=2
+            [pg.ready() for pg in more_pgs], timeout=20, num_returns=2
         )
         assert len(ready) == 2
